🚚 Move SayHi to patterns
