+++ src/pages/work.js
@@ -3,12 +3,13 @@ import { graphql } from "gatsby"
 
 import { PageLayout } from "../layouts"
 import SEO from "../components/SEO"
-import { SayHi } from "../components/SayHi/SayHi"
 import Link from "../components/Link"
 import { Box } from "../components/Box"
 import { Paragraph, Heading2 } from "../components/Typography"
 import List from "../components/List/List"
 import { Separator } from "../components/Separator"
+
+import { SayHi } from "../patterns"
 
 const Work = ({ data }) => {
   return (
