+++ src/patterns/index.ts
@@ -1,2 +1,3 @@
 export * from "./Footer/Footer"
 export * from "./Navigation/Navigation"
+export * from "./SayHi/SayHi"
